+++ src/Debug/ServiceInfo.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Neusta\ConverterBundle\Dump;
+namespace Neusta\ConverterBundle\Debug;
 
 /**
  * @internal
